requirements.txt
# install from https://pytorch.org/get-started/locally/
# torch>=1.12.0
# torchvision>=0.6.1
demucs>=4.0.1
pytube>=15.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.6